Add jQuery.postcodify
Postcodify is a jQuery plugin and server-side API for searching
South Korean street addresses and postal codes.

The Korean government lets you download a dataset of over 6 million
street addresses and associated information as a bunch of zip files.
The server-side API of Postcodify makes it easy to convert this information
into a MySQL database and search it in various ways.

The jQuery plugin helps you integrate this search engine into any web page,
such as the checkout page of an e-commerce website where the customer
must enter their shipping and billing addresses.

The jQuery plugin can be used either with the free public API
that the author operates, or with a search engine on your own machine.
About one hundred Korean websites are already using Postcodify.

Both the server-side API and the jQuery plugin component of Postcodify
are free software, released under LGPLv3.