Adding Library AmCharts
###AmCharts

URL: http://www.amcharts.com/javascript-charts/

The above libraries is free to use. Please check the section "Free even
for commercial use" on the above Library's URL

Since this library is not hosted on github, auto update has not been
configured.

The author for all the Library has been set to "AmCharts"