Adding Library AmStockCharts 3.13.0
### AmStockCharts

URL: http://www.amcharts.com/stock-chart/

The above libraries is free to use. Please check the section "Free even
for commercial use" on the above Library's URL

Since this library is not hosted on github, auto update has not been
configured.

The author for the Library has been set to "AmCharts"